8256740: ZGC: Move closures out of zOopClosure files
